Moved SMSDatabaseTest to android-test/database
After code review of Riccardo De Zen
