+++ lib/src/main/java/com/dezen/riccardo/smshandler/database/SMSDao.java
@@ -9,7 +9,7 @@ import com.dezen.riccardo.smshandler.SMSMessage;
  * Interface extending the BaseDao class for SMSMessage
  */
 @Dao
-public abstract class SMSDao extends BaseDao<SMSMessage>{
+abstract class SMSDao extends BaseDao<SMSMessage>{
     /**
      * @return the name of the table containing the SMSMessage entities.
      */
